requirements.txt
psycopg2-binary
dotenv
Flask